added null-test to method flushCache in file Downloader.scala
git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/sbaz/trunk@6887 5e8d7ff9-d8ef-0310-90f0-a4852d11357a

+++ src/sbaz/Downloader.scala
@@ -39,9 +39,11 @@ class Downloader(val dir:File) {
 
   // Delete all downloaded files
   def flushCache = {
-    for(val ent <- dir.listFiles().toList;
-        !ent.isDirectory()) {
-      ent.delete()
-    }
+    val files = dir.listFiles()
+    if (files != null)
+      for (val ent <- files.toList;
+           !ent.isDirectory()) {
+        ent.delete()
+      }
   }
 }
